fix: Update Watson Assistant code to allow for destruction of resource (#85)
BREAKING CHANGE: In this release, the code to manage Watson Assistant was updated to allow for deletion of the Assistant when the RAG sample DA is undeployed. As this is a different mechanism, when upgrading the DA to this or later versions, users may end up with a duplicate Assistant within their Watson Assistant instance. This should not incur additional costs as it will be unused, but we recommend that users navigate to their Watson Assistant instance after the upgrade, and delete the original Assistant from there. Prefixes for assistants are now available, so we recommend using a prefix to generate a unique name for the new assistant.